Reorganize the internals
