+++ blackjax/__init__.py
@@ -1,8 +1,13 @@
 from .diagnostics import effective_sample_size as ess
 from .diagnostics import potential_scale_reduction as rhat
-from .mcmc import hmc, nuts, rmh
-from .mcmc_adaptation import window_adaptation
-from .smc import adaptive_tempered_smc, tempered_smc
+from .kernels import (
+    hmc,
+    nuts,
+    rmh,
+    window_adaptation,
+    adaptive_tempered_smc,
+    tempered_smc,
+)
 
 __version__ = "0.3.0"
