+++ blackjax/adaptation/__init__.py
@@ -1,0 +1,1 @@
+from . import window_adaptation
