+++ blackjax/adaptation/step_size.py
@@ -5,7 +5,7 @@ import jax
 import jax.numpy as jnp
 
 import blackjax.optimizers as optimizers
-from blackjax.inference.hmc.hmc import HMCState
+from blackjax.mcmc.hmc import HMCState
 
 __all__ = [
     "dual_averaging_adaptation",
