+++ blackjax/smc/__init__.py
@@ -1,0 +1,4 @@
+from . import adaptive_tempered
+from . import tempered
+
+__all__ = ["adaptive_tempered", "tempered"]
