+++ blackjax/smc/solver.py
