+++ tests/test_solver.py
@@ -7,7 +7,7 @@ import jax.numpy as jnp
 import numpy as np
 from absl.testing import absltest, parameterized
 
-import blackjax.inference.smc.solver as solver
+import blackjax.smc.solver as solver
 
 solving_methods_to_test = [solver.dichotomy]
